restructure: remove entires directory
