--- src/entries/web-compiler.js
@@ -1,4 +1,0 @@
-/* @flow */
-
-export { parseComponent } from 'sfc/parser'
-export { compile, compileToFunctions } from 'web/compiler/index'
