--- src/entries/weex-framework.js
@@ -1,1 +1,0 @@
-export * from 'weex/framework'
