+++ src/platforms/web/compiler.js
@@ -1,0 +1,4 @@
+/* @flow */
+
+export { parseComponent } from 'sfc/parser'
+export { compile, compileToFunctions } from './compiler/index'
