+++ src/platforms/web/runtime.js
@@ -1,0 +1,5 @@
+/* @flow */
+
+import Vue from './runtime/index'
+
+export default Vue
