+++ src/platforms/weex/compiler.js
@@ -1,3 +1,1 @@
-/* @flow */
-
 export { compile } from 'weex/compiler/index'
